move IsChangeStreamOptionsMinRequiredMajorVersion
